+++ chain.go
@@ -249,13 +249,16 @@ func (b *jobStreamBuilder) joinArm() joinArmSpec {
 	if b == nil {
 		return joinArmSpec{}
 	}
-	spec := joinArmSpec{chain: b, region: b.region}
+	spec := joinArmSpec{region: b.region}
 	if b.job != nil {
 		spec.chainErr = b.job.err
 		if len(b.job.inputs) == 1 {
 			spec.chainInput = b.job.inputs[0]
 			spec.chainInputOK = true
 		}
+	}
+	if b.stream != nil {
+		spec.chainOperations = cloneOperationSpecs(b.stream.operations)
 	}
 	return spec
 }